moved integration tests to its own module, and added Jacoco code coverage plugin
