+++ tests/src/test/java/com/oracle/weblogic/imagetool/tests/utils/ExecCommand.java
@@ -1,7 +1,7 @@
 // Copyright (c) 2019, 2020, Oracle Corporation and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package com.oracle.weblogic.imagetool.integration.utils;
+package com.oracle.weblogic.imagetool.tests.utils;
 
 import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
