+++ tests/src/test/java/com/oracle/weblogic/imagetool/tests/utils/ExecResult.java
@@ -1,7 +1,7 @@
 // Copyright (c) 2019, 2020, Oracle Corporation and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package com.oracle.weblogic.imagetool.integration.utils;
+package com.oracle.weblogic.imagetool.tests.utils;
 
 /**
  * Class that holds the results of using java to exec a command (i.e. exit value, stdout and stderr)
