+++ tests/src/test/resources/wdt/build-archive.sh
