+++ tests/src/test/resources/wdt/domain.properties
