+++ tests/src/test/resources/wdt/simple-app/WEB-INF/web.xml
