+++ tests/src/test/resources/wdt/simple-topology.yaml
