+++ tests/src/test/resources/wdt/simple-topology1.yaml
