+++ tests/src/test/resources/wdt/simple-topology2.yaml
